People: moved entity mapping to service using manual mapping

+++ MovieDatabase.Core/Services/PeopleService.cs
@@ -1,5 +1,5 @@
-﻿using AutoMapper;
-using Microsoft.AspNetCore.Http;
+﻿using FluentValidation;
+using MovieDatabase.Core.Mapping;
 using MovieDatabase.Core.Models;
 using MovieDatabase.Core.Models.Dto;
 using MovieDatabase.Core.Repository.IRepository;
@@ -10,58 +10,72 @@ namespace MovieDatabase.Core.Services
     public class PeopleService : IPeopleService
     {
         private readonly IPeopleRepository _peopleRepository;
-        private readonly IMapper _mapper;
+        private readonly IValidator<People> _peopleValidator;
 
-        public PeopleService(IPeopleRepository peopleRepository, IMapper mapper)
+        public PeopleService(IPeopleRepository peopleRepository, IValidator<People> peopleValidator)
         {
             _peopleRepository = peopleRepository;
-            _mapper = mapper;
+            _peopleValidator = peopleValidator;
         }
 
-        public async Task<IEnumerable<People>> GetAllAsync(Expression<Func<People, bool>>? filter = null, string? includeProperties = null)
+        public async Task<IEnumerable<PeopleDto>> GetAllAsync(Expression<Func<People, bool>>? filter = null, string? includeProperties = null)
         {
-            return await _peopleRepository.GetAllAsync(filter, includeProperties);
+            var peopleList = await _peopleRepository.GetAllAsync(filter, includeProperties);
+
+            return peopleList.Select(p => p.MapToResponse());
         }
 
-        public async Task<People?> GetByIdAsync(int id, bool tracked = true, string? includeProperties = null)
+        public async Task<PeopleDto?> GetByIdAsync(int id, bool tracked = true, string? includeProperties = null)
         {
-            return await _peopleRepository.GetByIdAsync(id, tracked, includeProperties);
+            var people = await _peopleRepository.GetByIdAsync(id, tracked, includeProperties);
+
+            if (people == null)
+                return null;
+
+            return people.MapToResponse();
         }
 
-        public async Task<IEnumerable<People?>> GetByNameAsync(string name)
+        public async Task<IEnumerable<PeopleDto>> GetByNameAsync(string name)
         {
             if (string.IsNullOrEmpty(name))
             {
-                return Enumerable.Empty<People>();
+                return Enumerable.Empty<PeopleDto>();
             }
 
-            return await _peopleRepository.GetByNameAsync(name);
+            var peopleList = await _peopleRepository.GetByNameAsync(name);
+            return peopleList.Select(p => p.MapToResponse());
         }
-        public async Task<People> CreateAsync(PeopleCreateDto peopleDto)
+        public async Task<PeopleDto> CreateAsync(PeopleCreateDto peopleDto)
         {
-            var people = _mapper.Map<People>(peopleDto);
+            var people = peopleDto.MapToPeople();
+            await _peopleValidator.ValidateAndThrowAsync(people);
+
             await _peopleRepository.CreateAsync(people);
-            return people;
+            return people.MapToResponse();
         }
 
-        public async Task DeleteAsync(People people)
+        public async Task<bool> DeleteAsync(int id)
         {
-            await _peopleRepository.RemoveAsync(people);
+            var people = await _peopleRepository.GetByIdAsync(id, tracked: true);
+
+            if (people is null)
+                return false;
+
+            return await _peopleRepository.RemoveAsync(people);
         }
 
-        public async Task<People> UpdateAsync(PeopleUpdateDto updatedPeople)
+        public async Task<PeopleDto?> UpdateAsync(PeopleUpdateDto updatedPeople)
         {
             var existingPerson = await _peopleRepository.GetByIdAsync(updatedPeople.Id, true);
 
             if (existingPerson == null)
-            {
-                throw new BadHttpRequestException($"Person with ID {updatedPeople.Id} not found.");
-            }
+                return null;
 
-            _mapper.Map(updatedPeople, existingPerson);
+            existingPerson = updatedPeople.MapToPeople(existingPerson);
+            await _peopleValidator.ValidateAndThrowAsync(existingPerson);
 
             await _peopleRepository.UpdateAsync(existingPerson);
-            return existingPerson;
+            return existingPerson.MapToResponse();
         }
     }
 }
